+++ Package.swift
@@ -12,13 +12,17 @@ let package = Package(
             targets: ["ContentBlockerConverter", "FilterEngine"]),
         .executable(
             name: "ConverterTool",
-            targets: ["CommandLineWrapper"])
+            targets: ["CommandLineWrapper"]),
+        .executable(
+            name: "FileLockTester",
+            targets: ["FileLockTester"])
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
         // .package(url: /* package url */, from: "1.0.0"),
         .package(url: "https://github.com/gumob/PunycodeSwift.git", .exact("3.0.0")),
         .package(url: "https://github.com/apple/swift-argument-parser", .exact("1.5.0")),
+        .package(url: "https://github.com/apple/swift-collections.git", .exact("1.1.4")),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -26,7 +30,7 @@ let package = Package(
         .executableTarget(
             name: "CommandLineWrapper",
             dependencies: [
-                "ContentBlockerConverter",
+                "FilterEngine",
                 .product(name: "ArgumentParser", package: "swift-argument-parser")
             ]),
         .target(
@@ -36,14 +40,19 @@ let package = Package(
             ]),
         .target(
             name: "FilterEngine",
-            dependencies: ["ContentBlockerConverter"]),
+            dependencies: [
+                "ContentBlockerConverter",
+                .product(name: "Collections", package: "swift-collections")
+            ]),
+        .executableTarget(
+            name: "FileLockTester",
+            dependencies: [ "FilterEngine" ]),
         .testTarget(
             name: "ContentBlockerConverterTests",
             dependencies: ["ContentBlockerConverter"],
             resources: [.copy("Resources/test-rules.txt")]),
         .testTarget(
             name: "FilterEngineTests",
-            dependencies: ["FilterEngine"],
-            resources: [.copy("Resources/advanced-rules.txt")])
+            dependencies: ["FilterEngine"])
     ]
 )